+++ repos/ktor/client/src/commonMain/kotlin/dev/inmo/micro_utils/repos/ktor/client/crud/KtorReadStandardCrudRepo.kt
@@ -47,4 +47,12 @@ class KtorReadStandardCrudRepo<ObjectType, IdType> (
         ),
         Boolean.serializer()
     )
+
+    override suspend fun count(): Long = client.uniget(
+        buildStandardUrl(
+            baseUrl,
+            countRouting
+        ),
+        Long.serializer()
+    )
 }