+++ repos/ktor/server/src/jvmMain/kotlin/dev/inmo/micro_utils/repos/ktor/server/key_value/KtorStandartWriteKeyValueRepo.kt
@@ -2,7 +2,7 @@ package dev.inmo.micro_utils.repos.ktor.server.key_value
 
 import dev.inmo.micro_utils.ktor.server.includeWebsocketHandling
 import dev.inmo.micro_utils.ktor.server.uniload
-import dev.inmo.micro_utils.repos.StandardWriteKeyValueRepo
+import dev.inmo.micro_utils.repos.WriteStandardKeyValueRepo
 import dev.inmo.micro_utils.repos.ktor.common.key_value.*
 import io.ktor.application.*
 import io.ktor.routing.*
@@ -10,7 +10,7 @@ import kotlinx.serialization.KSerializer
 import kotlinx.serialization.builtins.PairSerializer
 
 fun <K, V> Route.configureWriteStandartKeyValueRepoRoutes (
-    originalRepo: StandardWriteKeyValueRepo<K, V>,
+    originalRepo: WriteStandardKeyValueRepo<K, V>,
     keySerializer: KSerializer<K>,
     valueSerializer: KSerializer<V>,
 ) {